Clarify section 2.12
Previously, this section suggested that Delayed Assembly documents should be used for episode and encounter documents while On-Demand documents are for summary documents. 

Both types of documents could be used with either paradigm, and the selection should be based on the capabilities and expectations of all actors in the community. This proposal attempts to clarify that.